BFont: Minor code cleanup and autolocking for AppFontManager
* changed explicit locking to use Autolocker for gFontManager/fAppFontManager
in ServerApp, per comments in https://review.haiku-os.org/c/haiku/+/4790
* changed BFont::LoadFont (memory version) to use size_t for size and offset
* no functional changes

Reviewed-on: https://review.haiku-os.org/c/haiku/+/6003
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>

+++ src/kits/interface/Font.cpp
@@ -1478,15 +1478,15 @@ BFont::LoadFont(const char* path)
 
 
 status_t
-BFont::LoadFont(const area_id fontAreaID, uint32 size, uint32 offset)
+BFont::LoadFont(const area_id fontAreaID, size_t size, size_t offset)
 {
 	BPrivate::AppServerLink link;
 
 	link.StartMessage(AS_ADD_FONT_MEMORY);
 
 	link.Attach<int32>(fontAreaID);
-	link.Attach<uint32>(size);
-	link.Attach<uint32>(offset);
+	link.Attach<size_t>(size);
+	link.Attach<size_t>(offset);
 
 	status_t status = B_ERROR;
 	if (link.FlushWithReply(status) != B_OK || status != B_OK) {
